feat: adding RabbitMqUserManagement et consortes (#77)
This PR introduces the `config_managed` `AuthenticationStrategy` to the Amazon MQ for RabbitMQ brokers and follows the pattern of the Amazon MQ for ActiveMQ bokers in terms of *user management*. As such - we introduce `RabbitMqUserManagement` type and deprecate direct setting of `admin` credentials on the brokers.

The `config_managed` strategy allows for using [OAuth authentication and authorization for RabbitMQ brokers](https://docs.aws.amazon.com/amazon-mq/latest/developer-guide/oauth-tutorial.html).

Additionally, as in `projen` the default Lambda Runtime for auto-generated lambda functions is set to `nodejs22.x` we're removing explicit setting in the `.projenrc.ts` file.